refactor: Add getTotalAction function
This commit adds a new function called getTotalAction to the action.ts file in the src/core/shared/infrastructure directory. The function is responsible for making an API call to retrieve the total count of records for a given resource. It takes a resource parameter and returns a Promise containing the total count. This function will be used to display the total count in the UI.